move parser caller into an object, and remove the extra file
