+++ curation-migration-datasync/externalCaller/parser.ts
@@ -1,5 +1,35 @@
 import { backOff } from 'exponential-backoff';
-import { parserCaller } from './parserCaller';
+
+import config from '../config';
+import fetch from 'node-fetch';
+
+type ParserMetadataResponse = {
+  resolvedId: string;
+  domainId: string;
+};
+
+/**
+ * Fetch domainID and resolvedID from the parser service,
+ * using the URL.
+ * @param url the URL to fetch data for
+ * @returns ParserMetadataResponse
+ */
+export const parser = {
+  parserCaller: async (url: string): Promise<ParserMetadataResponse> => {
+    const params = new URLSearchParams({
+      output: 'regular',
+      getItem: '1',
+      images: '0',
+      url: url,
+    });
+    const res = await fetch(config.parserEndpoint + '/' + params.toString());
+    const jsonRes = await res.json();
+    return {
+      domainId: jsonRes['item']['domain_id'],
+      resolvedId: jsonRes['resolved_id'],
+    };
+  },
+};
 
 /**
  * calls parser to fetch domainID and resolvedId.
@@ -15,7 +45,7 @@ export async function getParserMetadata(url: string): Promise<any> {
   //backoff doesn't allow us to type the response, but its enforced in parserCaller()
   let res: any;
   try {
-    res = await backOff(() => parserCaller(url), backOffOptions);
+    res = await backOff(() => parser.parserCaller(url), backOffOptions);
     console.log(res);
   } catch (e) {
     throw new Error(e);
